+++ src/main/java/com/today/flower/store/StoreImgRepository.java
@@ -4,9 +4,9 @@ import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StoreImgRepository extends JpaRepository<StoreImg, Long> {
+public interface StoreImgRepository extends JpaRepository<StoreImg, Integer> {
 	
 	List<StoreImg> findByIdOrderByIdAsc(Integer storeId);
 
-	StoreImg findByIdAndRepimgYn(Long storeId, String repimgYn);
+	StoreImg findByIdAndRepimgYn(Integer storeId, String repimgYn);
 }
